All issue details that should be editable can now be edited/updated.
Added issue created on (DATETIME) and issue created by details.

Removed support for uploading files (was not working).

Changed from @RequestScoped to @ApplicationScoped in CategoryResource and PriorityResource classes, because there data do not have to be retrieved every time the new issue page is loaded.

Cleaned up some code in front and backend.